+++ src/main/java/chess/command/Move.java
@@ -1,4 +1,4 @@
-package chess.domain.command;
+package chess.command;
 
 import chess.domain.ChessGame;
 import chess.domain.position.Position;
